refactor:all id use Long type and use user Auth

+++ src/main/java/org/codeplay/playcoolbackend/dto/OrderRequestDto.java
@@ -10,9 +10,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder
 public class OrderRequestDto {
-    private Integer userId;
-    private Integer concertId;
-    private Integer seatId;
+    private Long userId;
+    private Long concertId;
+    private Long seatId;
     private Integer price;
     private String paymentMethod;
 }
